requirements.txt
Django==1.8.3
django-bootstrap3==6.1.0
djangorestframework==3.1.3
Pygments==2.0.2
wheel==0.24.0
